move WaitAndTail to cli package
